Fixed a bug in the Windows shell extension that made desktop icons flicker
Root cause: we were using SHChangeNotify(SHCNE_ASSOCCHANGED...) to refresh
the shell which is unnecessarily strong and refreshes everything. Switching
to SHChangeNotify(SHCNE_UPDATEDIR... refreshes only the root anchor.

That being said, there's an even deeper root cause which is that the
shellext is getting disconnected from the gui very often and then
reconnects automatically. Upon reconnection, we clear the cache, which was
triggering the refresh. It is still unclear why the shellext is getting
disconnected so often, but it has to do with Windows unloading it from
memory when its not needed.

+++ src/shellext/win_explorer/src/ShellExtension/string_helpers.h
@@ -119,7 +119,7 @@ inline bool str_ends_with(const std::wstring& str, const std::wstring& suffix)
   */
 inline std::string hexencode(std::string arg) {
 	std::stringstream ss;
-	for (int i = 0 ; i < arg.size() ;i++) {
+	for (size_t i = 0; i < arg.size(); i++) {
 		ss << std::hex << (unsigned int)((unsigned char)arg[i]);
 	}
 	return ss.str();
